Split old to new code
