+++ server/app.py
@@ -7,20 +7,21 @@ from flask import request, session, g
 from flask_restful import Resource
 from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
-from marshmallow import (
-    Schema,
-    fields,
-    validates,
-    ValidationError,
-    pre_load,
-    validates_schema,
-)
+from marshmallow import Schema, fields, validates, ValidationError, pre_load
 from marshmallow.validate import Length
+from schemas import (
+    UserSchema,
+    CharacterSchema,
+    CampaignSchema,
+    UserUpdateSchema,
+)  # , CharacterCampaignSchema
 
 # Local imports
 from config import app, db, api
+
 # Add your model imports
 from models import db, User, Character, Campaign, CharacterCampaign
+import ipdb
 
 # Views go here!
 
@@ -29,118 +30,19 @@ from models import db, User, Character, Campaign, CharacterCampaign
 #     return '<h1>Project Server</h1>'
 
 
-#! Schema's Marshmallow
-class UserSchema(Schema):
-    id = fields.Int(dump_only=True)
-    username = fields.Str(
-        required=True,
-        validate=Length(min=2),
-        metadata={"description": "The unique username of the user"},
-    )
-    password_hash = fields.Str(
-        load_only=True,
-        required=True,
-        metadata={"description": "The password of the user"},
-    )
-    email = fields.Str(metadata={"description": "The email of the user"}) #! NEED REGEX! 
-    game_master = fields.Boolean(metadata={"description": "The option to determine if this user is a game master or not"})
-
-    @validates("username")
-    def validate_username(self, value):
-        if self.context.get("is_signup"):
-            if get_one_by_condition(User, User.username == value):
-                raise ValidationError("Username already exists")
-        else:  # This is the login case
-            if not get_one_by_condition(User, User.username == value):
-                raise ValidationError("Username does not exist")
-
-    @validates("email")
-    def validate_email(self, value):
-        if self.context.get("is_signup") and get_one_by_condition(
-            User, User.email == value
-        ):
-            raise ValidationError("Email already exists")
-    
-    @validates_schema
-    def validate_password(self, data, **kwargs):
-        if self.context.get('require_password') and 'password_hash' not in data:
-            raise ValidationError('Password is required', 'password_hash')
-
-
-    @pre_load
-    def strip_strings(self, data, is_signup=None, **kwargs):
-        extra_data = kwargs.get("extra_data")
-        print(f"Extra data: {extra_data}")
-        #! example use of kwags:
-        #! user_schema.load(data, extra_data="extra")
-        #! can do something with this like logging or tracking things
-        for key, value in data.items():
-            if isinstance(value, str):
-                data[key] = value.strip()
-        return data
-
-class CampaignSchema(Schema):
-    id = fields.Int(dump_only=True)
-    name = fields.Str(
-        required=True,
-        validate=Length(min=2),
-        metadata={"description": "The unique name of the campaign"},
-    )
-    description = fields.Str(
-        metadata={"description": "The description of the campaign"}
-    )
-    gamemaster_id = fields.Int(
-        metadata={"description": "The ID of the game master of the campaign"}
-    )
-
-    @validates("name")
-    def validate_name(self, value):
-        if self.context.get("is_create"):
-            if get_one_by_condition(Campaign, Campaign.name == value):
-                raise ValidationError("Campaign name already exists")
-        else:  # This is the update case
-            if not get_one_by_condition(Campaign, Campaign.name == value):
-                raise ValidationError("Campaign name does not exist")
-
-    @pre_load
-    def strip_strings(self, data, **kwargs):
-        for key, value in data.items():
-            if isinstance(value, str):
-                data[key] = value.strip()
-        return data
-
-class CharacterSchema(Schema):
-    id = fields.Int(dump_only=True)
-    name = fields.Str(
-        required=True,
-        validate=Length(min=2),
-        metadata={"description": "The unique name of the character"},
-    )
-    class_ = fields.Str(metadata={"description": "The class of the character"})
-    race = fields.Str(metadata={"description": "The race of the character"})
-    alignment = fields.Str(metadata={"description": "The alignment of the character"})
-    age = fields.Int(metadata={"description": "The age of the character"})
-    alive = fields.Boolean(metadata={"description": "The alive status of the character"})
-    description = fields.Str(metadata={"description": "The description of the character"})
-    user_id = fields.Int(metadata={"description": "The user id associated with the character"})
-
-    @pre_load
-    def strip_strings(self, data, **kwargs):
-        for key, value in data.items():
-            if isinstance(value, str):
-                data[key] = value.strip()
-        return data
-
 #! helpers
 def execute_query(query):
     return db.session.execute(query).scalars()
+
 
 def get_all(model):
     # return db.session.execute(select(model)).scalars().all()
     return execute_query(select(model)).all()
 
+
 def get_instance_by_id(model, id):
     return db.session.get(model, id)
+
 
 def get_one_by_condition(model, condition):
     # stmt = select(model).where(condition)
@@ -148,62 +50,23 @@ def get_one_by_condition(model, condition):
     # return result.scalars().first()
     return execute_query(select(model).where(condition)).first()
 
+
 def get_all_by_condition(model, condition):
     # stmt = select(model).where(condition)
     # result = db.session.execute(stmt)
     # return result.scalars().all()
     return execute_query(select(model).where(condition)).all()
 
-def get_current_user():
-    user_id = session.get("user_id")
-    if user_id is not None:
-        return get_instance_by_id(User, user_id)
-    return None
-
-def get_current_character():
-    character_id = session.get("character_id")
-    if character_id is not None:
-        return get_instance_by_id(Character, character_id)
-    return None
-
-def get_current_campaign():
-    campaign_id = session.get("campaign_id")
-    if campaign_id is not None:
-        return get_instance_by_id(Campaign, campaign_id)
-    return None
-
 
 # ? before request - verify session login
 @app.before_request
 def load_logged_in_user():
-
-    def get_current_user():
-        user_id = session.get("user_id")
-        if user_id is not None:
-            return get_instance_by_id(User, user_id)
-        return None
-
-    def get_current_character():
-        character_id = session.get("character_id")
-        if character_id is not None:
-            return get_instance_by_id(Character, character_id)
-        return None
-
-    def get_current_campaign():
-        campaign_id = session.get("campaign_id")
-        if campaign_id is not None:
-            return get_instance_by_id(Campaign, campaign_id)
-        return None
-
-    g.user = get_current_user()
-    g.character = get_current_character()
-    g.campaign = get_current_campaign()
-    # user_id = session.get("user_id")
-    # if user_id is None:
-    #     g.user = None
-    # else:
-    #     g.user = get_instance_by_id(User, user_id)
-    # #! Refactor this, remove recipebyid + consider additional adds
+    user_id = session.get("user_id")
+    if user_id is None:
+        g.user = None
+    else:
+        g.user = get_instance_by_id(User, user_id)
+    #! Refactor this, remove recipebyid + consider additional adds
     # path_dict = {"userbyid": User, "recipebyid": Recipe}
 
     # # If the current request's endpoint is in the dictionary
@@ -308,6 +171,7 @@ class BaseResource(Resource):
             db.session.rollback()
             return {"message": "Invalid data"}, 422
 
+
 # ? User Account Signup/Login/Logout/Session Resources
 class Signup(Resource):
     model = User
@@ -336,6 +200,8 @@ class Signup(Resource):
         g.user = user
 
         return self.schema.dump(user), 201
+
+
 class CheckSession(Resource):
 
     def get(self):
@@ -350,6 +216,7 @@ class CheckSession(Resource):
             "bio": user.bio,
             "image_url": user.image_url,
         }, 200
+
 
 class Login(Resource):
     model = User
@@ -372,6 +239,7 @@ class Login(Resource):
         g.user = user
         return {"id": user.id, "username": user.username}, 200
 
+
 class Logout(Resource):
     def delete(self):
         if (user_id := session.get("user_id")) is None:
@@ -380,35 +248,40 @@ class Logout(Resource):
         session["username"] = None
         return {}, 204
 
+
 class CharacterIndex(BaseResource):
     model = Character
     schema = CharacterSchema()
 
     def get(self):
-        if g.user is None or g.character is None:
+        # if (user_id := session.get("user_id")) is None:
+        if g.user is None:
             return {"message": "Unauthorized"}, 401
         return super().get(condition=Character.user_id == g.user.id)
 
     def post(self):
-        if g.user is None or g.character is None:
+        # if (_ := session.get("user_id")) is None:
+        if g.user is None:
             return {"message": "Unauthorized"}, 401
-        self.schema.context = {"require_password": True}
         return super().post()
 
-    def delete(self, id):
-        if g.user is None or g.character is None:
+    def delete(self, character_id=None):
+        ipdb.set_trace()
+        if g.user is None:
             return {"message": "Unauthorized"}, 401
-        return super().delete(id)
+        if character_id is None:
+            character_id = g.user.id
+        return super().delete(g.user.id)
 
     def patch(self, id):
-        if g.user is None or g.character is None:
+        if g.user is None:
             return {"message": "Unauthorized"}, 401
-        self.schema.context = {"require_password": False}
         return super().patch(id)
+
 
 class UsersIndex(BaseResource):
     model = User
-    schema = UserSchema()
+    schema = UserUpdateSchema()
 
     def get(self):
         # if (user_id := session.get("user_id")) is None:
@@ -433,12 +306,15 @@ class UsersIndex(BaseResource):
             return {"message": "Unauthorized"}, 401
         if user_id is None:
             user_id = g.user.id
-        return super().delete(user_id)
+        return super().delete(g.user.id)
 
-    def patch(self, id):
+    def patch(self, user_id=None):
         if g.user is None:
             return {"message": "Unauthorized"}, 401
+        if user_id is None:
+            user_id = g.user.id
         return super().patch(g.user.id)
+
 
 class CampaignsIndex(BaseResource):
     model = Campaign
@@ -449,16 +325,16 @@ class CampaignsIndex(BaseResource):
             return {"message": "Unauthorized"}, 401
         if campaign_id is None:
             campaign_id = g.user.id
-        return super().get(condition=(Campaign.gamemaster_id == g.user.id))
+        return super().get(condition=(Campaign.id == campaign_id))
 
     def delete(self, campaign_id=None):
         if g.user is None:
             return {"message": "Unauthorized"}, 401
         if campaign_id is None:
-            campaign_id = g.campaign_id
+            campaign_id = g.user.id
+
         # Get the campaign
-        # campaign = Campaign.query.get(campaign_id)
-        campaign = get_one_by_condition(Campaign.gamemaster_id == g.user.id)
+        campaign = Campaign.query.get(campaign_id)
 
         # Check if the campaign exists
         if not campaign:
@@ -477,9 +353,6 @@ class CampaignsIndex(BaseResource):
     def patch(self, campaign_id):
         if g.user is None:
             return {"message": "Unauthorized"}, 401
-        campaign = get_one_by_condition(Campaign.gamemaster_id == g.user.id)
-        if not campaign:
-            return {"message": "Campaign not found"}, 404
         return super().patch(campaign_id)
 
 
@@ -490,10 +363,15 @@ api.add_resource(Logout, "/logout", endpoint="logout")
 
 
 api.add_resource(UsersIndex, "/profile", "/profile/<int:user_id>", endpoint="profile")
-api.add_resource(CharacterIndex, "/characters", endpoint="characters")
+api.add_resource(
+    CharacterIndex,
+    "/characters",
+    "/characters/<int:character_id>",
+    endpoint="characters",
+)
 api.add_resource(
     CampaignsIndex, "/campaigns", "/campaigns/<int:campaign_id>", endpoint="campaigns"
 )
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     app.run(port=5555, debug=True)
